requirements.txt
opencv-python==4.8.1.78
numpy==1.24.3
pytesseract==0.3.10
python-dotenv==1.0.0
Pillow==10.0.0